LPS-54222 move MonitoringFilter to module
